+++ crates/runner-nordic/src/board.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use wasefire_board_api::{self as board, AppletMemory, Event, Failure, Singleton};
+use wasefire_board_api::{self as board, Event, Singleton};
 use wasefire_scheduler as scheduler;
 
 use crate::{Board, with_state};
@@ -36,6 +36,8 @@ pub mod timer;
 #[cfg(feature = "uart")]
 pub mod uart;
 pub mod usb;
+#[cfg(feature = "_vendor")]
+mod vendor;
 
 impl board::Api for Board {
     fn try_event() -> Option<board::Event<Board>> {
@@ -48,18 +50,6 @@ impl board::Api for Board {
                 None => Events::wait(),
                 Some(event) => break event,
             }
-        }
-    }
-
-    fn vendor(mem: impl AppletMemory, x1: u32, x2: u32, x3: u32, x4: u32) -> Result<u32, Failure> {
-        #[cfg(not(feature = "test-vendor"))]
-        let _ = &mem;
-        match (x1, x2, x3, x4) {
-            #[cfg(feature = "test-vendor")]
-            (0, _, _, _) => syscall_test::process(mem, x2, x3, x4),
-            #[cfg(feature = "gpio")]
-            (0x80000000, x, y, z) => Ok(gpio::syscall(x, y, z)?),
-            _ => Err(Failure::TRAP),
         }
     }
 
@@ -84,6 +74,8 @@ impl board::Api for Board {
     type Uart = uart::Impl;
     #[cfg(feature = "_usb")]
     type Usb = usb::Impl;
+    #[cfg(feature = "_vendor")]
+    type Vendor = vendor::Impl;
 }
 
 impl Singleton for crate::storage::Storage {