+++ crates/scheduler/src/call/debug.rs
@@ -13,8 +13,9 @@
 // limitations under the License.
 
 use wasefire_applet_api::debug::{self as api, Api, Perf};
+use wasefire_board_api::applet::{Memory as _, MemoryExt as _};
 use wasefire_board_api::debug::Api as _;
-use wasefire_board_api::{self as board, Api as Board, AppletMemory as _, AppletMemoryExt as _};
+use wasefire_board_api::{self as board, Api as Board};
 
 use crate::{DispatchSchedulerCall, SchedulerCall, Trap};