+++ crates/scheduler/src/call/store/fragment.rs
@@ -17,7 +17,7 @@ use wasefire_applet_api::store::fragment::Api;
 use wasefire_applet_api::store::fragment::{self as api};
 use wasefire_board_api::Api as Board;
 #[cfg(feature = "board-api-storage")]
-use wasefire_board_api::{AppletMemory as _, AppletMemoryExt as _};
+use wasefire_board_api::applet::{Memory as _, MemoryExt as _};
 #[cfg(feature = "board-api-storage")]
 use wasefire_store::fragment;