+++ crates/protocol-usb/CHANGELOG.md
@@ -40,4 +40,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 5 -->
+<!-- Increment to skip CHANGELOG.md test: 6 -->